feat: implement lua redis.pcall
redis.pcall is the same as redis.call, but, if an internal error
of any kind happens, it returns 'a table', instead of failing.

It's unclear from the docs what this table is supposed to be.

The only place I've seen pcall used is in redlock, which ignores
the return value:

https://github.com/mike-marcacci/node-redlock/blob/e90e80a611c83a6a8c16912c7c5a8c6e9d49c617/redlock.js#L31

I didn't check this is sufficient for redlock to work, but it
seems like a step in the right direction.